+++ app/db.py
@@ -5,6 +5,10 @@ from tornado.options import options
 from operator import itemgetter
 from datetime import datetime
 
+import random
+import string
+import hashlib
+import time
 import logging
 import hashlib
 import re
@@ -104,25 +108,24 @@ class Entity():
                 created_by,
                 created
             ) SELECT /* SQL_NO_CACHE */
-                r.relationship_definition_keyname,
+                rr.relationship_definition_keyname,
                 %s,
-                r.related_entity_id,
+                rr.related_entity_id,
                 %s,
                 NOW()
-            FROM
-                relationship AS r
-            WHERE r.relationship_definition_keyname IN ('leecher', 'viewer', 'editor', 'owner')
-            AND r.deleted IS NULL
-            AND r.entity_id IN (
-                SELECT DISTINCT entity_id
-                FROM relationship
-                WHERE deleted IS NULL
-                AND related_entity_id = %s
-                AND relationship_definition_keyname = 'child'
-            );
+            FROM      relationship r
+            LEFT JOIN relationship rr ON rr.entity_id = r.entity_id
+            WHERE     r.deleted IS NULL
+            AND       r.related_entity_id = %s
+            AND       r.relationship_definition_keyname = 'child'
+            AND       rr.relationship_definition_keyname IN ('leecher', 'viewer', 'editor', 'owner' );
         """
         # logging.debug(sql)
         self.db.execute(sql, entity_id, self.created_by, entity_id)
+
+        # Populate default values
+        for default_value in self.db.query('SELECT keyname, defaultvalue FROM property_definition WHERE entity_definition_keyname = %s AND defaultvalue IS NOT null', entity_definition_keyname):
+            self.set_property(entity_id=entity_id, property_definition_keyname=default_value.keyname, value=default_value.defaultvalue)
 
         # Propagate properties
         sql = """
@@ -209,7 +212,7 @@ class Entity():
         elif definition.datatype == 'file':
             if value:
                 uploaded_file = value
-                value = self.db.execute_lastrowid('INSERT INTO file SET filename = %s, file = %s, created_by = %s, created = NOW();', uploaded_file['filename'], uploaded_file['body'], self.created_by)
+                value = self.db.execute_lastrowid('INSERT INTO file SET filename = %s, filesize = %s, file = %s, created_by = %s, created = NOW();', uploaded_file['filename'], len(uploaded_file['body']), uploaded_file['body'], self.created_by)
             field = 'value_file'
         elif definition.datatype == 'boolean':
             field = 'value_boolean'
@@ -242,6 +245,11 @@ class Entity():
                     value,
                     self.created_by
                 )
+
+        self.db.execute('UPDATE entity SET changed = NOW(), changed_by = %s WHERE id = %s;',
+            self.created_by,
+            entity_id,
+        )
 
         return new_property_id
 
@@ -545,6 +553,7 @@ class Entity():
                     entity_definition.%(language)s_label_plural     AS entity_label_plural,
                     entity_definition.%(language)s_description      AS entity_description,
                     entity.created                                  AS entity_created,
+                    entity.changed                                  AS entity_changed,
                     entity.public                                   AS entity_public,
                     entity_definition.%(language)s_displayname      AS entity_displayname,
                     entity_definition.%(language)s_displayinfo      AS entity_displayinfo,
@@ -572,6 +581,7 @@ class Entity():
                     property.value_datetime                         AS value_datetime,
                     property.value_entity                           AS value_entity,
                     property.value_counter                          AS value_counter,
+                    property.value_reference                        AS value_reference,
                     property.value_file                             AS value_file
                 FROM
                     entity,
@@ -604,7 +614,8 @@ class Entity():
                 items.setdefault('item_%s' % row.entity_id, {})['description'] = row.entity_description
                 items.setdefault('item_%s' % row.entity_id, {})['sort'] = row.entity_sort
                 items.setdefault('item_%s' % row.entity_id, {})['sort_value'] = row.entity_sort_value
-                items.setdefault('item_%s' % row.entity_id, {})['created'] = formatDatetime(row.entity_created, '%(day)02d.%(month)02d.%(year)d') if row.entity_created else ''
+                items.setdefault('item_%s' % row.entity_id, {})['created'] = row.entity_created
+                items.setdefault('item_%s' % row.entity_id, {})['changed'] = row.entity_changed
                 items.setdefault('item_%s' % row.entity_id, {})['displayname'] = row.entity_displayname
                 items.setdefault('item_%s' % row.entity_id, {})['displayinfo'] = row.entity_displayinfo
                 items.setdefault('item_%s' % row.entity_id, {})['displaytable'] = row.entity_displaytable
@@ -645,8 +656,13 @@ class Entity():
                     db_value = row.value_datetime
                     value = formatDatetime(row.value_datetime)
                 elif row.property_datatype == 'reference':
+                    value = ''
+                    if row.value_reference:
+                        reference = self.__get_properties(entity_id=row.value_reference)
+                        if reference:
+                            value = reference[0].get('displayname')
+                    logging.debug(str(reference))
                     db_value = row.value_entity
-                    value = row.value_entity
                 elif row.property_datatype == 'file':
                     db_value = row.value_file
                     blobstore = self.db.get('SELECT id, filename, filesize FROM file WHERE id=%s LIMIT 1', row.value_file)
@@ -923,6 +939,9 @@ class Entity():
 
         """
 
+        if type(file_id) is not list:
+            file_id = [file_id]
+
         if self.user_id:
             public = ''
         else:
@@ -931,6 +950,7 @@ class Entity():
         sql = """
             SELECT
                 f.id,
+                f.created,
                 f.file,
                 f.filename
             FROM
@@ -938,22 +958,15 @@ class Entity():
                 property AS p,
                 property_definition AS pd
             WHERE p.value_file = f.id
-            AND property_definition.keyname = p.property_definition_keyname
-            AND f.id = %(file_id)s
+            AND pd.keyname = p.property_definition_keyname
+            AND f.id IN (%(file_id)s)
             %(public)s
             AND p.deleted IS NULL
-            LIMIT 1
-            """ % {'file_id': file_id, 'public': public}
+            """ % {'file_id': ','.join(map(str, file_id)), 'public': public}
         # logging.debug(sql)
 
-        result = self.db.get(sql)
+        return self.db.query(sql)
 
-        if not result:
-            return
-        if not result.file:
-            return
-
-        return result
 
     def get_entity_definition(self, entity_definition_keyname):
         """
@@ -1124,18 +1137,19 @@ class User():
                 user.language,
                 user.email,
                 user.picture,
-                user_profile.provider
+                user.provider
             FROM
                 property_definition,
                 property,
-                user,
-                user_profile
+                entity,
+                user
             WHERE property.property_definition_keyname = property_definition.keyname
+            AND entity.id = property.entity_id
             AND property.deleted IS NULL
+            AND entity.deleted IS NULL
             AND user.email = property.value_string
-            AND user_profile.user_id = user.id
             AND property_definition.dataproperty = 'user'
-            AND user_profile.session = %s
+            AND user.session = %s
             LIMIT 1;
         """, session)
 
@@ -1151,34 +1165,34 @@ class User():
     def __getitem__(self, key):
         return getattr(self, key)
 
-    def create(self, provider='', id='', email='', name='', picture='', language='', session=''):
+    def login(self, request_handler, session_key=None, provider=None, provider_id=None, email=None, name=None, picture=None):
         """
-        Creates new (or updates old) user.
+        Starts session. Creates new (or updates old) user.
 
         """
+        if not session_key:
+            session_key = str(''.join(random.choice(string.ascii_letters + string.digits) for x in range(32)) + hashlib.md5(str(time.time())).hexdigest())
+        user_key = hashlib.md5(request_handler.request.remote_ip + request_handler.request.headers.get('User-Agent', None)).hexdigest()
+
+
         db = connection()
-        profile_id = db.execute_lastrowid('INSERT INTO user_profile SET provider = %s, provider_id = %s, email = %s, name = %s, picture = %s, session = %s, created = NOW() ON DUPLICATE KEY UPDATE email = %s, name = %s, picture = %s, session = %s, changed = NOW();',
+        profile_id = db.execute_lastrowid('INSERT INTO user SET provider = %s, provider_id = %s, email = %s, name = %s, picture = %s, language = %s, session = %s, created = NOW() ON DUPLICATE KEY UPDATE email = %s, name = %s, picture = %s, session = %s, changed = NOW();',
                 provider,
-                id,
+                provider_id,
                 email,
                 name,
                 picture,
-                session,
+                request_handler.settings['default_language'],
+                session_key+user_key,
                 email,
                 name,
                 picture,
-                session
+                session_key+user_key
             )
-        profile = db.get('SELECT id, user_id FROM user_profile WHERE id = %s', profile_id)
 
-        if not profile.user_id:
-            user_id = db.execute_lastrowid('INSERT INTO user SET email = %s, name = %s, picture = %s, language = %s, created = NOW();',
-                email,
-                name,
-                picture,
-                language
-            )
-            db.execute('UPDATE user_profile SET user_id = %s WHERE id = %s;', user_id, profile.id)
+        request_handler.set_secure_cookie('session', session_key)
+
+        return session_key
 
 
 def formatDatetime(date, format='%(day)02d.%(month)02d.%(year)d %(hour)02d:%(minute)02d'):